requirements.txt
pandas
numpy
matplotlib
seaborn
jupyter
python-dotenv
openai
langchain
langchain-openai
